We're now building libcc with autotools.
Now to build the actual tasbot.
WIP

More WIP!

Ok, cc-lib seems to pass most of the tests...

Strangely this is despite the fact that I'm not actually building
everything in it. I doubt that finishing that will magically make
this test pass, but who knows?

Either way, getting the rest of cc-lib to build shouldn't be too
hard, and then it's just a matter of getting tasbot built.

Oops, no need for that makefile

Yeah, that's the one we want

Needed to add pthreads, works great now

imore wip

Ok, it builds!

Building this whatever midithing

Still need libsdl as a dependency

Ok, just needed to include all that sdl stuff