Ok, cc-lib seems to pass most of the tests...
Strangely this is despite the fact that I'm not actually building
everything in it. I doubt that finishing that will magically make
this test pass, but who knows?

Either way, getting the rest of cc-lib to build shouldn't be too
hard, and then it's just a matter of getting tasbot built.